ステージの見た目を変更した。それからコンポーネントの複製が正常でなかったのでコピー代入演算子を実装した。あと削除コマンドを実装した

+++ mtgb/FbxAnimationController.cpp
@@ -82,11 +82,11 @@ void mtgb::FbxAnimationController::UpdateFrame()
     {
         if (isLooping_)
         {
-            currentFrame_ = endFrame;
+            currentFrame_ = static_cast<float>(endFrame);
         }
         else
         {
-            currentFrame_ = startFrame;
+            currentFrame_ = static_cast<float>(startFrame);
             isPlaying_ = false;
             isFinished_ = true;
         }
@@ -95,11 +95,11 @@ void mtgb::FbxAnimationController::UpdateFrame()
     {
         if (isLooping_)
         {
-            currentFrame_ = startFrame;
+            currentFrame_ = static_cast<float>(startFrame);
         }
         else
         {
-            currentFrame_ = endFrame;
+            currentFrame_ = static_cast<float>(endFrame);
             isPlaying_ = false;
             isFinished_ = true;
         }
